draft: extract shared CubeSetupPanel from DraftPage (#1253)
Move the inline CubeSetupPanel + NumberField + DEFAULT_CUBE_SETTINGS
out of DraftPage.tsx into a new shared component at
`client/src/components/draft/CubeSetupPanel.tsx`. The shared module
has zero store imports — it takes a required `onStart` callback
that receives the assembled `{ cubeName, cubeListText, settings }`
payload.

DraftPage now passes an onStart closure that grabs `difficulty` and
`startCubeDraft` from the solo draft store. DraftPodPage (Commit 7)
will pass a different onStart that writes to draftPodStore.cubeForm.

N3: the panel composes an external `disabled` prop with its internal
loading state via OR, so a busy parent can suppress submission
without racing against the panel's own loading lock.

Pure refactor — solo cube draft regression coverage is unchanged.
