+++ client/src/pages/DraftPage.tsx
@@ -6,6 +6,7 @@ import { useNavigate, useSearchParams } from "react-router";
 import { useDraftStore } from "../stores/draftStore";
 import { CardPreview } from "../components/card/CardPreview";
 import type { CardHoverInfo } from "../components/card/CardPreview";
+import { CubeSetupPanel } from "../components/draft/CubeSetupPanel";
 import { DraftIntro } from "../components/draft/DraftIntro";
 import { DraftSteps } from "../components/draft/DraftSteps";
 import { SetSelector } from "../components/draft/SetSelector";
@@ -15,10 +16,8 @@ import { DraftProgress } from "../components/draft/DraftProgress";
 import { LimitedDeckBuilder } from "../components/draft/LimitedDeckBuilder";
 import { ScreenChrome } from "../components/chrome/ScreenChrome";
 import { menuButtonClass } from "../components/menu/buttonStyles";
-import { fetchCubeList } from "../services/cubeCobra";
 import { runLimits } from "../services/quickDraftPersistence";
 import type { DraftRunFormat, DraftRunState } from "../services/quickDraftPersistence";
-import type { CubeDraftSettings } from "../adapter/draft-adapter";
 import { usePreferencesStore } from "../stores/preferencesStore";
 
 // ── Format Picker ─────────────────────────────────────────────────────
@@ -30,194 +29,6 @@ const FORMAT_OPTIONS: Array<{ value: DraftRunFormat; labelKey: string; descKey: 
 ];
 
 type DraftSetupMode = "set" | "cube";
-
-const DEFAULT_CUBE_SETTINGS: CubeDraftSettings = {
-  pod_size: 8,
-  pack_count: 3,
-  cards_per_pack: 15,
-  min_deck_size: 40,
-  addable_cards: {
-    policy: "StandardBasics",
-    custom: [],
-  },
-};
-
-function CubeSetupPanel() {
-  const { t } = useTranslation("draft");
-  const [cubeName, setCubeName] = useState(t("cubeSetup.defaultCubeName"));
-  const [cubeText, setCubeText] = useState("");
-  const [cubeUrl, setCubeUrl] = useState("");
-  const [settings, setSettings] = useState<CubeDraftSettings>(DEFAULT_CUBE_SETTINGS);
-  const [customAddables, setCustomAddables] = useState("");
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-
-  const updateSetting = (key: keyof Omit<CubeDraftSettings, "addable_cards">, value: number) => {
-    setSettings((prev) => ({ ...prev, [key]: value }));
-  };
-
-  const updateCustomAddables = (value: string) => {
-    setCustomAddables(value);
-    setSettings((prev) => ({
-      ...prev,
-      addable_cards: {
-        ...prev.addable_cards,
-        custom: value
-          .split("\n")
-          .map((line) => line.trim())
-          .filter(Boolean),
-      },
-    }));
-  };
-
-  const handleFetchUrl = async () => {
-    if (!cubeUrl.trim()) return;
-    setLoading(true);
-    setError(null);
-    try {
-      setCubeText(await fetchCubeList(cubeUrl));
-    } catch (err) {
-      setError(err instanceof Error ? err.message : t("cubeSetup.fetchError"));
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const handleStart = async () => {
-    setLoading(true);
-    setError(null);
-    try {
-      const { difficulty, startCubeDraft } = useDraftStore.getState();
-      await startCubeDraft(cubeText, cubeName, settings, difficulty);
-    } catch (err) {
-      setError(err instanceof Error ? err.message : t("cubeSetup.startError"));
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const canStart = cubeText.trim().length > 0 && !loading;
-
-  return (
-    <div className="flex flex-col gap-4">
-      <div className="grid gap-3 md:grid-cols-[1fr_220px_220px_220px]">
-        <label className="flex flex-col gap-1">
-          <span className="text-xs uppercase tracking-[0.16em] text-white/35">{t("cubeSetup.cubeName")}</span>
-          <input
-            value={cubeName}
-            onChange={(e) => setCubeName(e.target.value)}
-            className="rounded-lg border border-white/10 bg-black/30 px-3 py-2 text-sm text-white outline-none focus:border-emerald-400/50"
-          />
-        </label>
-        <NumberField label={t("cubeSetup.seats")} value={settings.pod_size} min={2} max={16} onChange={(v) => updateSetting("pod_size", v)} />
-        <NumberField label={t("cubeSetup.packs")} value={settings.pack_count} min={1} max={6} onChange={(v) => updateSetting("pack_count", v)} />
-        <NumberField label={t("cubeSetup.packSize")} value={settings.cards_per_pack} min={1} max={30} onChange={(v) => updateSetting("cards_per_pack", v)} />
-      </div>
-
-      <div className="grid gap-3 md:grid-cols-[220px_1fr_auto]">
-        <NumberField label={t("cubeSetup.minDeck")} value={settings.min_deck_size} min={1} max={100} onChange={(v) => updateSetting("min_deck_size", v)} />
-        <label className="flex flex-col gap-1">
-          <span className="text-xs uppercase tracking-[0.16em] text-white/35">{t("cubeSetup.exportUrl")}</span>
-          <input
-            value={cubeUrl}
-            onChange={(e) => setCubeUrl(e.target.value)}
-            placeholder={t("cubeSetup.exportUrlPlaceholder")}
-            className="rounded-lg border border-white/10 bg-black/30 px-3 py-2 text-sm text-white outline-none placeholder:text-white/25 focus:border-emerald-400/50"
-          />
-        </label>
-        <button
-          type="button"
-          onClick={handleFetchUrl}
-          disabled={loading || !cubeUrl.trim()}
-          className={menuButtonClass({ tone: "neutral", size: "md", disabled: loading || !cubeUrl.trim(), className: "self-end" })}
-        >
-          {t("cubeSetup.loadUrl")}
-        </button>
-      </div>
-
-      <div className="grid gap-3 md:grid-cols-[260px_1fr]">
-        <label className="flex flex-col gap-1">
-          <span className="text-xs uppercase tracking-[0.16em] text-white/35">{t("cubeSetup.deckAddables")}</span>
-          <select
-            value={settings.addable_cards.policy}
-            onChange={(e) =>
-              setSettings((prev) => ({
-                ...prev,
-                addable_cards: {
-                  ...prev.addable_cards,
-                  policy: e.target.value as CubeDraftSettings["addable_cards"]["policy"],
-                },
-              }))
-            }
-            className="rounded-lg border border-white/10 bg-black/30 px-3 py-2 text-sm text-white outline-none focus:border-emerald-400/50"
-          >
-            <option value="StandardBasics">{t("cubeSetup.addablesStandardBasics")}</option>
-            <option value="StandardBasicsPlusCustom">{t("cubeSetup.addablesBasicsPlusCustom")}</option>
-            <option value="CustomOnly">{t("cubeSetup.addablesCustomOnly")}</option>
-          </select>
-        </label>
-        <label className="flex flex-col gap-1">
-          <span className="text-xs uppercase tracking-[0.16em] text-white/35">{t("cubeSetup.customAddableCards")}</span>
-          <textarea
-            value={customAddables}
-            onChange={(e) => updateCustomAddables(e.target.value)}
-            placeholder={t("cubeSetup.customAddablePlaceholder")}
-            className="min-h-10 resize-y rounded-lg border border-white/10 bg-black/30 px-3 py-2 text-sm text-white outline-none placeholder:text-white/25 focus:border-emerald-400/50"
-          />
-        </label>
-      </div>
-
-      <textarea
-        value={cubeText}
-        onChange={(e) => setCubeText(e.target.value)}
-        spellCheck={false}
-        placeholder="1 Lightning Bolt&#10;1 Black Lotus&#10;1 Tropical Island"
-        className="min-h-[280px] resize-y rounded-lg border border-white/10 bg-black/35 p-3 font-mono text-sm leading-6 text-white outline-none placeholder:text-white/25 focus:border-emerald-400/50"
-      />
-
-      {error && <div className="rounded-lg border border-red-400/30 bg-red-500/10 px-3 py-2 text-sm text-red-200">{error}</div>}
-
-      <div className="flex justify-end">
-        <button
-          type="button"
-          onClick={handleStart}
-          disabled={!canStart}
-          className={menuButtonClass({ tone: "emerald", size: "lg", disabled: !canStart })}
-        >
-          {t("cubeSetup.startCubeDraft")}
-        </button>
-      </div>
-    </div>
-  );
-}
-
-function NumberField({
-  label,
-  value,
-  min,
-  max,
-  onChange,
-}: {
-  label: string;
-  value: number;
-  min: number;
-  max: number;
-  onChange: (value: number) => void;
-}) {
-  return (
-    <label className="flex flex-col gap-1">
-      <span className="text-xs uppercase tracking-[0.16em] text-white/35">{label}</span>
-      <input
-        type="number"
-        min={min}
-        max={max}
-        value={value}
-        onChange={(e) => onChange(Math.min(max, Math.max(min, Number(e.target.value))))}
-        className="rounded-lg border border-white/10 bg-black/30 px-3 py-2 text-sm text-white outline-none focus:border-emerald-400/50"
-      />
-    </label>
-  );
-}
 
 function FormatPicker({ onLaunch }: { onLaunch: () => void }) {
   const { t } = useTranslation("draft");
@@ -607,7 +418,12 @@ export function DraftPage() {
             {setupMode === "set" ? (
               <SetSelector onStartDraft={handleStartDraft} />
             ) : (
-              <CubeSetupPanel />
+              <CubeSetupPanel
+                onStart={async ({ cubeName, cubeListText, settings }) => {
+                  const { difficulty, startCubeDraft } = useDraftStore.getState();
+                  await startCubeDraft(cubeListText, cubeName, settings, difficulty);
+                }}
+              />
             )}
           </div>
         )}
